pyproject.toml
[project]
name = "landgraph-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.74",
    "langchain-core>=0.2.38",
    "langsmith>=0.1.63",
    "langchain-anthropic>=0.2.0",
    "tavily-python>=0.3.0",
    "langchain-google-genai>=2.1.10",
    "python-dotenv>=1.1.1",
    "pytz>=2024.1",
    "mcp-use>=1.3.10",
]
